mypage 모듈화
